+++ lib/src/widgets/text_line.dart
@@ -17,6 +17,7 @@ import '../models/documents/nodes/line.dart';
 import '../models/documents/nodes/node.dart';
 import '../models/documents/style.dart';
 import '../utils/color.dart';
+import '../utils/font.dart';
 import '../utils/platform.dart';
 import 'box.dart';
 import 'controller.dart';
@@ -307,7 +308,7 @@ class _TextLineState extends State<TextLine> {
         if (k == Attribute.underline.key || k == Attribute.strikeThrough.key) {
           var textColor = defaultStyles.color;
           if (color?.value is String) {
-            textColor = stringToColor(color?.value);
+            textColor = stringToColor(color?.value, textColor);
           }
           res = _merge(res.copyWith(decorationColor: textColor),
               s!.copyWith(decorationColor: textColor));
@@ -342,19 +343,7 @@ class _TextLineState extends State<TextLine> {
           res = res.merge(defaultStyles.sizeHuge);
           break;
         default:
-          double? fontSize;
-          if (size.value is double) {
-            fontSize = size.value;
-          } else if (size.value is int) {
-            fontSize = size.value.toDouble();
-          } else if (size.value is String) {
-            fontSize = double.tryParse(size.value);
-          }
-          if (fontSize != null) {
-            res = res.merge(TextStyle(fontSize: fontSize));
-          } else {
-            throw 'Invalid size ${size.value}';
-          }
+          res = res.merge(TextStyle(fontSize: getFontSize(size.value)));
       }
     }
 
@@ -394,7 +383,7 @@ class _TextLineState extends State<TextLine> {
   }
 
   Future<void> _launchUrl(String url) async {
-    await launch(url);
+    await launchUrl(Uri.parse(url));
   }
 
   void _tapNodeLink(Node node) {